Change all database ids to Long
add transaction to insert new emails to MailboxDao
Move MailboxActivity, SignInActivity, SignUpActivity to appropriate
packages.

+++ src/main/kotlin/com/email/scenes/ActivityMessage.kt
@@ -7,5 +7,5 @@ import com.email.scenes.composer.data.ComposerInputData
  */
 
 sealed class ActivityMessage {
-    data class SendMail(val emailId: Int, val composerInputData: ComposerInputData): ActivityMessage()
+    data class SendMail(val emailId: Long, val composerInputData: ComposerInputData): ActivityMessage()
 }
